Pull Request: Validações nos Cruds de Usuário e Estoque, estilização do sistema e relatório

+++ src/main/java/edu/unialfa/alberguepro/service/UsuarioService.java
@@ -3,6 +3,7 @@ package edu.unialfa.alberguepro.service;
 import edu.unialfa.alberguepro.model.Usuario;
 import edu.unialfa.alberguepro.repository.UsuarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,13 +18,13 @@ public class UsuarioService {
     private PasswordEncoder passwordEncoder;
 
     public void salvar(Usuario usuario) {
-        // Se o usuário já existe
+        // validacao se o usuario já existe
         if (usuario.getId() != null) {
-            // Se uma nova senha foi fornecida (não está em branco), criptografa e atualiza
+            // se uma nova senha foi fornecida (não está em branco), criptografa e atualiza
             if (usuario.getPassword() != null && !usuario.getPassword().isEmpty()) {
                 usuario.setPassword(passwordEncoder.encode(usuario.getPassword()));
             } else {
-                // Senão, busca a senha atual no banco e a mantém
+                // se não, busca a senha atual no banco e a mantém
                 usuarioRepository.findById(usuario.getId()).ifPresent(usuarioExistente -> {
                     usuario.setPassword(usuarioExistente.getPassword());
                 });
@@ -36,11 +37,20 @@ public class UsuarioService {
         usuarioRepository.save(usuario);
     }
 
-        public void excluir(Long id) {
+    public void excluir(Long id) {
+        String usernameLogado = SecurityContextHolder.getContext().getAuthentication().getName();
+
+        Usuario usuarioParaExcluir = usuarioRepository.findById(id)
+                .orElseThrow(() -> new IllegalStateException("Usuário não encontrado: " + id));
+
+        if (usuarioParaExcluir.getUsername().equals(usernameLogado)) {
+            throw new IllegalStateException("Não é possível excluir o usuário logado.");
+        }
+
         usuarioRepository.deleteById(id);
     }
 
-     @Transactional
+@Transactional
     public void toggleAtivo(Long id) {
         // Busca o usuário no banco ou lança uma exceção se não encontrar
         Usuario usuario = usuarioRepository.findById(id)
